Revert "fix test"
This reverts commit e3af13f2b1325cc8ca562a1725f76521207b746e.
