+++ packages/cardpay-reward-api/tests/test_reward_api.py
@@ -3,7 +3,7 @@
 import pytest
 from cardpay_reward_api.database import Base, get_db
 from cardpay_reward_api.indexer import Indexer
-from cardpay_reward_api.main import app, get_db
+from cardpay_reward_api.main import app
 from cardpay_reward_api.models import Root
 from fastapi.testclient import TestClient
